+++ server/middleware/twofa.go
@@ -2,17 +2,13 @@ package middleware
 
 import (
 	"context"
+	"encoding/json"
+	"errors"
 	"fmt"
+	"io/ioutil"
 	"log"
-
 	"strings"
 
-	"io/ioutil"
-
-	"encoding/json"
-
-	"github.com/mcuadros/go-version"
-	"github.com/pkg/errors"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
 	"gopkg.in/alecthomas/kingpin.v2"
@@ -55,19 +51,6 @@ func NewTwoFA(object, operation, id string, codes []string) *twoFA {
 	}
 }
 
-func getVersionId(ctx context.Context) (string, error) {
-	headers, ok := metadata.FromIncomingContext(ctx)
-	if !ok {
-		log.Println("Cannot get header metadata from context")
-		return "", errors.New("Cannot get header metadata from context")
-	}
-
-	if headers["version"] == nil || len(headers["version"]) == 0 {
-		return "", errors.New("Version not found in the header")
-	}
-	return headers["version"][0], nil
-}
-
 func getTwoFA(ctx context.Context) (*twoFA, error) {
 	md, ok := metadata.FromIncomingContext(ctx)
 	if !ok {
@@ -89,13 +72,6 @@ func getTwoFA(ctx context.Context) (*twoFA, error) {
 	}
 
 	return NewTwoFA(md["2faobject"][0], md["2faoperation"][0], md["2faident"][0], strings.Split(md["2facodes"][0], ",")), nil
-}
-
-// Check the version of the client's binary.
-// Return false, if version is deprecated.
-func validateVersion(cliVersion, leastVersion string) bool {
-	c := version.NewConstrainGroupFromString(">=" + leastVersion)
-	return c.Match(cliVersion)
 }
 
 func contains(items []string, match string) bool {
@@ -158,42 +134,5 @@ func TwoFAInterceptor() grpc.UnaryServerInterceptor {
 
 		// call
 		return nil, nil
-	}
-}
-
-func UnaryServerInterceptor(supportVersion string) grpc.UnaryServerInterceptor {
-	return func(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
-		url := "https://github.com/tsocial/tessellate/releases"
-
-		// Get the version from the header.
-		version, err := getVersionId(ctx)
-		if err != nil {
-			return nil, err
-		}
-
-		versionErr := errors.Errorf(
-			"You are using an older version: %v of Tessellate CLI. "+
-				"Download the newer version (>= %v) from: %v",
-			version, supportVersion, url)
-
-		// If the id is empty, return a older version error.
-		if version == "" {
-			log.Printf("Version not found.")
-			return nil, versionErr
-		}
-
-		if !validateVersion(version, supportVersion) {
-			return nil, versionErr
-		}
-
-		// Else, pass the request ahead to the handler.
-		resp, err := handler(ctx, req)
-		if err != nil {
-			fmt.Printf("%+v", err)
-			return nil, err
-		}
-
-		// Return handler's response and err.
-		return resp, nil
 	}
 }
